refactor(teradata): use format_destination_name pattern instead of ClassVar flag
Replace the `sanitize_destination_name: ClassVar[bool]` flag on the base
`SQLUploader` with an overridable `format_destination_name` method on
`TeradataUploader`, matching the established convention used by AstraDB,
Pinecone, and Weaviate connectors.

The `_sanitize_table_name` helper is moved from `sql.py` to `teradata.py`
since sanitization is purely a Teradata concern. The base class no longer
carries any sanitization concept — the flag was misleading because no base
class method ever read it.

`create_destination` now follows the AstraDB pattern: sanitize the
auto-generated `destination_name` first via `format_destination_name`,
then prefer the user-provided `upload_config.table_name` unchanged.
